requirements.txt
fastapi
uvicorn
httpx
Pillow
transformers
#torch==2.1.0
accelerate
bitsandbytes
python-multipart
sentence_transformers

# Paste this shit:
# pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118